requirements.txt
werkzeug==0.15.4 # needed for airflow 1.10.3
apache-airflow[gcp,kubernetes,slack]==1.10.9
